requirements.txt
# 必要なPythonパッケージ一覧
# このファイルを使用してインストール: pip install -r requirements.txt

# ==================================================
# 深層学習フレームワーク
# ==================================================
torch>=2.0.0              # PyTorch - GPU推論に必要
torchvision>=0.15.0       # 画像処理とモデルサポート

# ==================================================
# コンピュータビジョン
# ==================================================
opencv-python>=4.8.0      # 動画処理・画像処理
ultralytics>=8.0.0        # YOLOv8モデル
supervision>=0.16.0       # トラッキングとアノテーション

# ==================================================
# 画像処理
# ==================================================
Pillow>=10.0.0            # 画像の読み込み・保存
numpy>=1.24.0             # 数値計算

# ==================================================
# GUI
# ==================================================
PySide6>=6.5.0            # Qt6ベースのGUIフレームワーク

# ==================================================
# プログレスバー・ユーティリティ
# ==================================================
tqdm>=4.65.0              # プログレスバー表示

# ==================================================
# 注意事項
# ==================================================
# ● TensorRT最適化（オプション）:
#   - NVIDIA GPU使用時にTensorRTで高速化可能
#   - TensorRTは別途インストールが必要（CUDA環境）
#   - インストール方法: https://developer.nvidia.com/tensorrt
#
# ● CUDA対応PyTorch（オプション）:
#   - GPU使用時は以下のコマンドで最新のCUDA対応版をインストール推奨
#   - pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#   - または: https://pytorch.org/ で環境に合わせたコマンドを確認
#
# ● 推奨Python バージョン:
#   - Python 3.8 以上
#   - Python 3.10 または 3.11 を推奨
